% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{ABC}
\alias{ABC}
\title{Robin Eriksson 2018
Selection of parameter posterior estimation methods.
--------------------------------------------------------------
Algortihm:
While  n < Nstop
--- t <- propose parameters
--- d <- Simulate data(t)
--- s <- transform into summary statistics(d)
--- if( || s_org - s || < epsilon )
--- --- store t in theta
--- --- n <- n + 1
--- else
--- --- n <- n
return theta
--------------------------------------------------------------
ABC estimator. Will perfome
Approximate Bayesian Computations in order
to approximate the posterior distribution
of the model parameters.}
\usage{
ABC(observation, Simulator, SummaryStatistics, Proposal, extraArgsEstimator,
  extraArgsSimulator, extraArgsSummaryStatistics, extraArgsProposal)
}
\arguments{
\item{observation}{Observed data, that parameters are to be fit to.}

\item{Simulator}{(function) Model simulator. Theta dependent.}

\item{SummaryStatistics}{(function) transformer of the output data from the simulator. Output is a vector.}

\item{Proposal}{proposal functions for the parameters.}

\item{extraArgsSimulator}{extra arguments for Simulator}

\item{extraArgsSummaryStatistics}{extra arguments for summary statistics transformer.}

\item{extraArgsProposal}{extra arguments for the proposal function}

\item{extraArgsABC}{extra argruments for ABC: epsilon, nStop, debug}
}
\value{
data frame with samples from the posterior distribution and their "ABC" distance.
}
\description{
Robin Eriksson 2018
Selection of parameter posterior estimation methods.
--------------------------------------------------------------
Algortihm:
While  n < Nstop
--- t <- propose parameters
--- d <- Simulate data(t)
--- s <- transform into summary statistics(d)
--- if( || s_org - s || < epsilon )
--- --- store t in theta
--- --- n <- n + 1
--- else
--- --- n <- n
return theta
--------------------------------------------------------------
ABC estimator. Will perfome
Approximate Bayesian Computations in order
to approximate the posterior distribution
of the model parameters.
}
