% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{abcGrid}
\alias{abcGrid}
\title{ABC grid exploration}
\usage{
abcGrid(observation, Simulator, SummaryStatistics, Proposal, extraArgsEstimator,
  extraArgsSimulator, extraArgsSummaryStatistics)
}
\arguments{
\item{observation}{Observed data, that parameters are to be fit to.}

\item{Simulator}{(function) Model simulator. Theta dependent.}

\item{SummaryStatistics}{(function) transformer of the output data from the simulator. Output is a vector.}

\item{extraArgsEstimator}{extra argruments for the estimator: nStop, nSim, debug}

\item{extraArgsSimulator}{extra arguments for Simulator}

\item{extraArgsSummaryStatistics}{extra arguments for summary statistics transformer.}

\item{proposal}{null}
}
\value{
grid parameters with distance
}
\description{
The method employs ABC distance measurements
on a grid around a center-point.
}
