% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{agg}
\alias{agg}
\title{aggregate the time series from the simulated data, also over quarters!}
\usage{
agg(data, extraArgs)
}
\arguments{
\item{data}{simulated data frame}

\item{extraArgs}{column in which data i found}
}
\description{
aggregate the time series from the simulated data, also over quarters!
}
