% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{aggWilkinson}
\alias{aggWilkinson}
\title{Siminf summary statistics
input should be
## simulation data
## nodes to sample
## time stamps to sample at (measurements)
output should be
### vector where each element is a summary statistic}
\usage{
aggWilkinson(data, extraArgs)
}
\arguments{
\item{data}{simulated data frame}

\item{extraArgs}{column in which data i found}
}
\description{
Different "main" functions here should act at different
vectors returned, i.e, different summary statistic combinations.
Could be usefull if we use different models, as SS choice is
often heuristic.
}
\details{
Robin Eriksson @ 2017
aggregate the time series from the simulated data.
and extract the Wilkinson statistics
}
