% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{bootstrap}
\alias{bootstrap}
\title{Naive bootstrap draw new samples!}
\usage{
bootstrap(mat, extraArgs)
}
\arguments{
\item{mat}{the data matrix}

\item{B}{the number of bootstrap samples}
}
\description{
Naive bootstrap draw new samples!
}
