% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{computeDelta}
\alias{computeDelta}
\title{Simulate data and return the synthetic log likelihood}
\usage{
computeDelta(Simulator, SummaryStatistics, extraArgsSimulator,
  extraArgsSummaryStatistics, distance, theta, ss_obs)
}
\arguments{
\item{Simulator}{the simulator}

\item{SummaryStatistics}{the summary statistics computer}

\item{extraArgsSimulator}{the arguments for the simulator}

\item{extraArgsSummaryStatistics}{the arguments for the statistics}

\item{theta}{the model parameter}

\item{ss_obs}{the observed summary statistics}
}
\value{
the synthetic likelihood of the proposed parameter theta
}
\description{
Simulate data and return the synthetic log likelihood
}
