% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{computeSL}
\alias{computeSL}
\title{Simulate data and return the synthetic log likelihood}
\usage{
computeSL(Simulator, SummaryStatistics, extraArgsSimulator,
  extraArgsSummaryStatistics, extraArgsEstimator, theta, ss_obs)
}
\arguments{
\item{Simulator}{the simulator}

\item{SummaryStatistics}{the summary statistics computer}

\item{extraArgsSimulator}{the arguments for the simulator}

\item{extraArgsSummaryStatistics}{the arguments for the statistics}

\item{extraArgsEstimator}{the arguments for the estimator}

\item{theta}{the model parameter}

\item{ss_obs}{the observed summary statistics}

\item{nSim}{the number of simulations}

\item{normalize}{if the summary statistics are to be normalized or not.}
}
\value{
the synthetic likelihood of the proposed parameter theta
}
\description{
Simulate data and return the synthetic log likelihood
}
