% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{computeWeights}
\alias{computeWeights}
\title{Compute the weights used in the weighted}
\usage{
computeWeights(data)
}
\arguments{
\item{data}{Observation data in matrix form}
}
\description{
Compute the weights used in the weighted
}
