% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findLast}
\alias{findLast}
\title{Helper function for timeMatch.
Find the last entry in the vector of the same entry as
in the "starting" position.}
\usage{
findLast(obs, start, node)
}
\description{
Helper function for timeMatch.
Find the last entry in the vector of the same entry as
in the "starting" position.
}
