% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceClass.R
\docType{data}
\name{Inference}
\alias{Inference}
\title{This file holds the (R6) inference class for
parameters estimation and posterior exploration.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Inference
}
\description{
An inference class that will hold the functionality of all the functions.
}
\keyword{datasets}
