% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initmodel.R
\name{init_model_widgren}
\alias{init_model_widgren}
\title{Implementation by Stefan Widgren with alterations
Init the SISe Model}
\usage{
init_model_widgren(model, theta, tspan, p = NULL, phi = NULL)
}
\description{
Every 1/p-th animal is moved to infected state.
}
