% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{maxSL}
\alias{maxSL}
\title{Numerically find maximumlikelihood estimate}
\usage{
maxSL(observation, Simulator, SummaryStatistics, Proposal, extraArgsEstimator,
  extraArgsSimulator, extraArgsSummaryStatistics, extraArgsProposal)
}
\arguments{
\item{observation}{Observed data, that parameters are to be fit to.}

\item{Simulator}{(function) Model simulator. Theta dependent.}

\item{SummaryStatistics}{(function) transformer of the output data from the simulator. Output is a vector.}

\item{Proposal}{proposal functions for the parameters.}

\item{extraArgsEstimator}{extra argruments for the estimator: nStop, nSim, debug}

\item{extraArgsSimulator}{extra arguments for Simulator}

\item{extraArgsSummaryStatistics}{extra arguments for summary statistics transformer.}

\item{extraArgsProposal}{extra arguments for the proposal function}
}
\value{
samples from the posterior distribution.
}
\description{
Numerically find maximumlikelihood estimate
}
