% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot2and3}
\alias{plot2and3}
\title{plot posterior in 2d and 3d side by side}
\usage{
plot2and3(data, invert = TRUE, ex = FALSE)
}
\arguments{
\item{data}{distance matrix data}

\item{invert}{invert the distance/likelihood}

\item{ex}{exponentiate the distance/likelihood}
}
\description{
plot
}
