% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot2dotherStuff}
\alias{plot2dotherStuff}
\title{plot posterior in 2d, with interpolated values.}
\usage{
plot2dotherStuff(data, type = "exp", theta0 = NULL, rec = c(0.0095, 0.09,
  0.0105, 0.1))
}
\arguments{
\item{data}{distance matrix data}

\item{invert}{invert the distance/likelihood}

\item{ex}{exponentiate the distance/likelihood}
}
\description{
plot
}
