% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot3dStuff}
\alias{plot3dStuff}
\title{plot posterior in 3d}
\usage{
plot3dStuff(data, invert = FALSE, ex = TRUE)
}
\arguments{
\item{data}{distance matrix data}

\item{invert}{invert the distance/likelihood}

\item{ex}{exponentiate the distance/likelihood}
}
\description{
plot
}
