% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotAggregate_mean}
\alias{plotAggregate_mean}
\title{Plot mean aggregated data}
\usage{
plotAggregate_mean(data, column = "sample")
}
\arguments{
\item{data}{distance matrix data}

\item{column}{which column to "aggregate"}
}
\value{
the aggregates (mean) time series.
}
\description{
Plot aggregated the data that is used when computing SS
}
