% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotAggregate_sum_qtr}
\alias{plotAggregate_sum_qtr}
\title{Plot mean aggregated data}
\usage{
plotAggregate_sum_qtr(listOfData, column = "sample")
}
\arguments{
\item{column}{which column to "aggregate"}

\item{data}{distance matrix data}
}
\value{
the aggregates (mean) time series.
}
\description{
Plot aggregated the data that is used when computing SS
}
