% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{pool_prevalence}
\alias{pool_prevalence}
\title{Computes the pool prevalence}
\usage{
pool_prevalence(S_n, I_n)
}
\arguments{
\item{S_n}{Number of susceptible individuals}

\item{I_n}{Mumber of infected individuals}
}
\value{
Pool prevalence
}
\description{
Computes the pool prevalence
}
