% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{predict_env_sample}
\alias{predict_env_sample}
\title{Predict outcome from a overshoe and pooled pat sample in a
simulated herd}
\usage{
predict_env_sample(wgpp, whpp)
}
\arguments{
\item{wgpp}{Vector with the within-group pool prevalence}

\item{whpp}{Vector with the within-herd pool prevalence}
}
\value{
logical vector
}
\description{
There must be at least one pool positive in the herd (whpp > 0)
Parameters from: Widgren, S.; Eriksson, E.; Aspan, A.; Emanuelson,
U.; Alenius, S. & Lindberg, A. Environmental sampling for
evaluating verotoxigenic Escherichia coli O157: H7 status in dairy
cattle herds. Journal of Veterinary Diagnostic Investigation
}
