% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{predict_env_sample_single}
\alias{predict_env_sample_single}
\title{like the above, but trying to emulate only using 1 group per herd.}
\usage{
predict_env_sample_single(whpp)
}
\arguments{
\item{whpp}{withing herd prevalence}
}
\description{
like the above, but trying to emulate only using 1 group per herd.
}
