% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{predict_env_sample_SISe}
\alias{predict_env_sample_SISe}
\title{Given the within herd prevalence determine if the node is pos or neg.}
\usage{
predict_env_sample_SISe(whpp)
}
\arguments{
\item{whpp}{withing herd prevalence}
}
\description{
Given the within herd prevalence determine if the node is pos or neg.
}
