% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predict_env_sample_SISe_Rcpp}
\alias{predict_env_sample_SISe_Rcpp}
\title{Predict the outcome of the env. sample}
\usage{
predict_env_sample_SISe_Rcpp(whpp)
}
\description{
Predict the outcome of the env. sample
}
