% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proposals.R
\name{Proposal_kernel_SISe_random_lognormal}
\alias{Proposal_kernel_SISe_random_lognormal}
\title{Proposal functions for all parameters in the model
Kernel function. Takes a value and alters it with a normal random walk.}
\usage{
Proposal_kernel_SISe_random_lognormal(theta, extraArgs)
}
\value{
names list with names lists for each parameter.
}
\description{
Proposal functions for all parameters in the model
Kernel function. Takes a value and alters it with a normal random walk.
}
