% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proposals.R
\name{Proposal_SISe3sp_fix}
\alias{Proposal_SISe3sp_fix}
\title{Proposal functions for all parameters in the model}
\usage{
Proposal_SISe3sp_fix()
}
\value{
names list with names lists for each parameter.
}
\description{
Proposal functions for all parameters in the model
}
