% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proposals.R
\name{Proposal_SISe_allUniformNarrow}
\alias{Proposal_SISe_allUniformNarrow}
\title{Proposal functions for all parameters in the SISe_sp model}
\usage{
Proposal_SISe_allUniformNarrow(theta, extraArgs)
}
\arguments{
\item{theta}{the old parameter vector}

\item{extraArgs}{not in use.}
}
\value{
proposed parameter vector
}
\description{
Propose a new parameter set. Independent of the lastly proposed
and in all dimensions at once.
}
