% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proposals.R
\name{Proposal_SISe_Kernel}
\alias{Proposal_SISe_Kernel}
\title{Proposal functions for all parameters in the model
Kernel function. Takes a value and alters it with a log-normal random walk.}
\usage{
Proposal_SISe_Kernel(S = 0.3)
}
\value{
names list with names lists for each parameter.
}
\description{
Proposal functions for all parameters in the model
Kernel function. Takes a value and alters it with a log-normal random walk.
}
