% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Proposals.R
\name{Proposal_SISe_real}
\alias{Proposal_SISe_real}
\title{Proposal functions for all parameters in the SISe_sp model}
\usage{
Proposal_SISe_real()
}
\value{
names list with names lists for each parameter.
}
\description{
Proposal functions for all parameters in the SISe_sp model
}
