% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{qtrStore}
\alias{qtrStore}
\title{Store data in quarters.}
\usage{
qtrStore(indata, timeCol, dataCol, logical = FALSE)
}
\arguments{
\item{data}{as list}
}
\value{
a data.frame from the data in the list.
}
\description{
Store data in quarters.
}
