% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{sample_nodes_par_Rcpp}
\alias{sample_nodes_par_Rcpp}
\title{Simulate environmental sampling in the nodes (in parallel!)}
\usage{
sample_nodes_par_Rcpp(result, cl0 = NULL)
}
\arguments{
\item{result}{The result from simulating the model, list of dataframes for each nodes to sample.}

\item{cl0}{the number of cores (threads) to use.}
}
\value{
sampled results. Positive or negative at the time vectors.
}
\description{
Simulate environmental sampling in the nodes (in parallel!)
}
