% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_nodes_Rcpp_seed}
\alias{sample_nodes_Rcpp_seed}
\title{A wrapped version of the node sampler including a seed}
\usage{
sample_nodes_Rcpp_seed(resMat)
}
\description{
A wrapped version of the node sampler including a seed
}
