% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{sample_pools}
\alias{sample_pools}
\title{Sample pools. Group individuals three and three in pools.
Randomly determine the status of each pool based on the number of
positive samples in each pool and the sensitivity.}
\usage{
sample_pools(S_n, I_n)
}
\arguments{
\item{S_n}{Number of susceptible individuals}

\item{I_n}{Mumber of infected individuals}
}
\value{
data.frame with the number of positive pools (pos) and the
    total number of pools (n).
}
\description{
Sample pools. Group individuals three and three in pools.
Randomly determine the status of each pool based on the number of
positive samples in each pool and the sensitivity.
}
