% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodeSampler.R
\name{sample_single_node}
\alias{sample_single_node}
\title{Sample one node given a vector of S and I values.}
\usage{
sample_single_node(Svec, Ivec)
}
\description{
Sample one node given a vector of S and I values.
}
