% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulators.R
\name{setTheta}
\alias{setTheta}
\title{Re-format theta to model format}
\usage{
setTheta(theta)
}
\arguments{
\item{theta}{named parameter vector}
}
\value{
named parameter vector of correct model format.
}
\description{
Re-format theta to model format
}
