% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimInfInference.R
\docType{package}
\name{SimInfInference}
\alias{SimInfInference}
\alias{SimInfInference-package}
\title{This package extends SimInf, by using it in a context
of Bayesian Parameter Inference}
\description{
This package extends SimInf, by using it in a context
of Bayesian Parameter Inference
}
