% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulators.R
\name{SimInfSimulator_sandbox}
\alias{SimInfSimulator_sandbox}
\title{This file holds all Simulator functions.
Robinit Eriksson @2017
SimInf included data (model)}
\usage{
SimInfSimulator_sandbox(theta, extraArgs)
}
\arguments{
\item{theta}{named vector of model parameters}

\item{extraArgs}{(runSeed, threads, nObs, tObs, tspan, includeTrue)
runSeed the seed used in the actual simulation
threads number of threads used, if null all available are used
nObs the nodes that are observed durring measurements. (needs to be a subset of the avaiable nodes)
tObs at what times that the nodes are observed. (needs to be a subset of tspan)
tspan time (integer) vector used in simulator. When the state has to be recorded.
includeTrue true if the underlying true state should be returned aswell.}
}
\value{
result matrix in a list. Columns are nodes, and rows are timestamps.
}
\description{
This file holds all Simulator functions.
Robinit Eriksson @2017
SimInf included data (model)
}
