% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{SLAM}
\alias{SLAM}
\title{--------------------------------------------------------------
SLAM (Synthetic likelihood Adaptive Metropolis  estimator.
Computes the synthetic likelihood (SL) (Wood 2010) and the performes a Metropolis-Hasting algortihm
with SL instead of the true likelihood}
\usage{
SLAM(observation, Simulator, SummaryStatistics, Proposal, extraArgsEstimator,
  extraArgsSimulator, extraArgsSummaryStatistics, extraArgsProposal)
}
\arguments{
\item{observation}{Observed data, that parameters are to be fit to.}

\item{Simulator}{(function) Model simulator. Theta dependent.}

\item{SummaryStatistics}{(function) transformer of the output data from the simulator. Output is a vector.}

\item{Proposal}{proposal functions for the parameters.}

\item{extraArgsEstimator}{extra argruments for the estimator: nStop, nSim, debug}

\item{extraArgsSimulator}{extra arguments for Simulator}

\item{extraArgsSummaryStatistics}{extra arguments for summary statistics transformer.}

\item{extraArgsProposal}{extra arguments for the proposal function}
}
\value{
samples from the posterior distribution.
}
\description{
--------------------------------------------------------------
SLAM (Synthetic likelihood Adaptive Metropolis  estimator.
Computes the synthetic likelihood (SL) (Wood 2010) and the performes a Metropolis-Hasting algortihm
with SL instead of the true likelihood
}
