% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{SLAMreInit}
\alias{SLAMreInit}
\title{helper function for SLAM when reinit the Markov chain.}
\usage{
SLAMreInit(observation, Simulator, SummaryStatistics, Proposal,
  extraArgsEstimator, extraArgsSimulator, extraArgsSummaryStatistics,
  extraArgsProposal)
}
\description{
helper function for SLAM when reinit the Markov chain.
}
