% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{SyntheticLogLikelihood}
\alias{SyntheticLogLikelihood}
\title{Compute the synthetic likelihood}
\usage{
SyntheticLogLikelihood(ss_obs, ss)
}
\arguments{
\item{ss_obs}{the observed synthetic likelihood}

\item{ss}{simulated summary statistics as a matrix, each row is a simulaton.}
}
\description{
Compute the synthetic likelihood
}
