% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimators.R
\name{SyntheticLogLikelihood_wood}
\alias{SyntheticLogLikelihood_wood}
\title{Synthetic likelihood as written in mfasilio @Github
This method replicates the proposed alterations defined
in the paper by Wood (Nature 2010)}
\usage{
SyntheticLogLikelihood_wood(y, X)
}
\arguments{
\item{y}{observation}

\item{X}{simulations}
}
\value{
synthetic log-likelihood
}
\description{
Synthetic likelihood as written in mfasilio @Github
This method replicates the proposed alterations defined
in the paper by Wood (Nature 2010)
}
