% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{theme_Publication}
\alias{theme_Publication}
\title{theme:ing for ggplots
Credit goes to:
Koundinya Desiraju (04/07/2015)}
\usage{
theme_Publication(base_size = 14)
}
\description{
theme:ing for ggplots
Credit goes to:
Koundinya Desiraju (04/07/2015)
}
