% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{weighted.var}
\alias{weighted.var}
\title{Weighted Variance
Credit to Dr. Gavin Simpson +44 (0)20 7679 0522}
\usage{
weighted.var(x, w, na.rm = FALSE)
}
\description{
Weighted Variance
Credit to Dr. Gavin Simpson +44 (0)20 7679 0522
}
