% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{weightedStat}
\alias{weightedStat}
\title{Compute mean and variance (weighted)}
\usage{
weightedStat(vec, w)
}
\arguments{
\item{vec}{summarized data in vector (as sum or mean from matrix)}

\item{w}{the weights}
}
\description{
Compute mean and variance (weighted)
}
