% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStatistics.R
\name{wilkinsonStatistics}
\alias{wilkinsonStatistics}
\title{Compute autoregression constants, mean, and log of variance}
\usage{
wilkinsonStatistics(vec, w = NULL)
}
\arguments{
\item{vec}{a named list with lists of vectors to extract summary statistics from}

\item{w}{weigts}
}
\description{
Compute autoregression constants, mean, and log of variance
}
